+++ api/sql/schema/340-rule.splitNameCustom.sql
@@ -1,0 +1,5 @@
+CREATE TYPE [rule].[splitNameCustom] AS TABLE (
+    conditionName NVARCHAR(100),
+    [name] VARCHAR(50),
+    tag VARCHAR(255)
+)