+++ api/sql/schema/450-rule.falsePropertyFactorCount.sql
@@ -16,7 +16,8 @@ BEGIN
             WHERE
                 p.factor IN ('so', 'do', 'co') AND
                 ct.conditionId = @conditionId
-            UNION SELECT
+            UNION ALL
+            SELECT
                 p.factor
             FROM
                 @properties p
@@ -26,6 +27,16 @@ BEGIN
                 [rule].conditionProperty ct ON ct.name = t.name AND ct.value = t.value
             WHERE
                 p.factor IN ('ss', 'ds', 'cs', 'oc', 'sc', 'dc') AND
+                ct.conditionId = @conditionId
+            UNION ALL
+            SELECT
+                p.factor
+            FROM
+                @properties p
+            JOIN
+                [rule].conditionProperty ct ON ct.name = p.name AND ct.value = p.value
+            WHERE
+                p.factor IN ('sk', 'st', 'dk', 'dt') AND
                 ct.conditionId = @conditionId
         )
     )