+++ config.js
@@ -56,6 +56,7 @@ module.exports = () => ({
         adapter: true
     },
     microservice: {
+        sqlTest: true,
         sql: {
             utc: true
         },
@@ -73,6 +74,12 @@ module.exports = () => ({
             exclude: joi.any(),
             clearing: joi.boolean()
         }).required().messages(utc),
+        sqlUnitTest: [
+            joi.boolean(),
+            joi.object({
+                exclude: joi.any()
+            })
+        ],
         sqlTest: [
             joi.boolean(),
             joi.object({