+++ test/unit.js
@@ -1,0 +1,4 @@
+const path = require('path');
+module.exports = function sqlUnitTest({config}) {
+    return config && {seed: [{path: path.join(__dirname, 'unit'), config}]};
+};