+++ ui/react/configuration/actions.js
@@ -3,6 +3,6 @@ import * as actionTypes from './actionTypes';
 export function setConfig(config) {
     return {
         type: actionTypes.SET_RULE_CONFIG,
-        config: config
+        config
     };
 }