+++ ui/react/pages/RuleProfile/Tabs/Split/Info/index.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import Input from 'ut-front-react/components/Input';
 import MultiSelectBubble from 'ut-front-react/components/MultiSelectBubble';
-import { splitTags } from '../../../../../../../common';
+import { splitTags } from '../../../../../../common';
 import {fromJS} from 'immutable';
 import style from '../../style.css';